// front/src/providers/api-client.generated/model/getInterviewsResponse.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InterviewDto } from './interviewDto';


export interface GetInterviewsResponse { 
    success: boolean;
    message?: string;
    error?: object;
    statusCode?: number;
    errorGuid?: string;
    token?: string;
    filteredResults: number;
    interviews: Array<InterviewDto>;
}