// front/src/providers/api-client.generated/model/interviewDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './userDto';
import { CandidateDto } from './candidateDto';


export interface InterviewDto { 
    id?: string;
    creationDate?: Date;
    modifDate?: Date;
    candidateId?: string;
    candidate?: CandidateDto;
    date?: Date;
    title?: string;
    comment?: string;
    consultantId?: string;
    consultant?: UserDto;
    guid?: string;
    candidateResponse?: string;
    agencyPlace?: string;
    /**
     * Whether the candidate did not show up to the interview
     */
    noShow?: boolean;
}